Data processing

AnnData object

scVI.AnnData — Type

mutable struct AnnData

Minimal Julia implementation of the Python AnnData object (see package documentation and Github repository).

Keyword arguments

countmatrix::Union{Matrix,Nothing}=nothing: countmatrix in cell x gene shape

ncells::Union{Int,Nothing}=nothing: number of cells; size(countmatrix,1)

ngenes::Union{Int,Nothing}=nothing: number of genes; size(countmatrix,2)

layers::Union{Dict,Nothing}=nothing: dictionary of other layers (e.g., normalized counts), corresponds to adata.layers

obs::Union{Dict,Nothing}=nothing: dictionary of information about cells, e.g., celltypes

summary_stats::Union{Dict, Nothing}=nothing: dictionary of summary information, corresponds to adata.uns["_scvi"]["summary_stats"]

registry::Union{Dict, Nothing}=nothing: dictionary corresponding to adata.uns["_scvi"]["data_registry"]

vars::Union{Dict, Nothing}=nothing: dictionary of information about genes/features, e.g., gene names or highly variable genes

celltypes=nothing: vector of cell type names, shorthand for adata.obs["cell_type"]

train_inds=nothing: vector of cell indices used for training an scVAE model

dataloader=nothing: Flux.DataLoader object used for training an scVAE model

scVI_latent=nothing: latent representation of trained scVAE model

scVI_latent_umap=nothing: UMAP of latent representation from trained scVAE model

is_trained::Bool=false: indicating whether a model has been trained on the AnnData object

Example

julia> adata = load_tasic("scvi/data/")
    AnnData object with a countmatrix with 1679 cells and 15119 genes
        layers dict with the following keys: ["normalized_counts", "counts"]
        unique celltypes: ["Vip", "L4", "L2/3", "L2", "Pvalb", "Ndnf", "L5a", "SMC", "Astro", "L5", "Micro", "Endo", "Sst", "L6b", "Sncg", "Igtp", "Oligo", "Smad3", "OPC", "L5b", "L6a"]
        training status: not trained

Library size and normalization

scVI.init_library_size — Method

init_library_size(adata::AnnData)

Computes and returns library size based on AnnData object.

Based on the scvi-tools function from here

Returns a tupe of arrays of length equal to the number of batches in adata as stored in adata.obs["batch_indices"], containing the means and variances of the library size in each batch in adata. If no batch indices are given, defaults to 1 batch.

scVI.estimate_size_factors — Method

estimate_size_factors(mat; locfunc=median)

Estimates size factors to use for normalization, based on the corresponding Seurat functionality. Assumes a countmatrix mat in cell x gene format as input, returns a vector of size factors.

For details, please see the Seurat documentation.

scVI.normalize_counts — Method

normalize_counts(mat::Abstractmatrix)

Normalizes the countdata in mat by dividing it by the size factors calculated with estimate_size_factors. Assumes a countmatrix mat in cell x gene format as input, returns the normalized matrix.

scVI.normalize_counts! — Method

normalize_counts(adata::AnnData)

Normalizes the adata.countmatrix by dividing it by the size factors calculated with estimate_size_factors. Adds the normalized count matrix to adata.layers and returns adata.

Highly variable genes

scVI.highly_variable_genes — Method

highly_variable_genes(adata::AnnData;
    layer::Union{String,Nothing} = nothing,
    n_top_genes::Int=2000,
    batch_key::Union{String,Nothing} = nothing,
    span::Float64=0.3
    )

Computes highly variable genes according to the workflows on scanpy and Seurat v3 per batch and returns a dictionary with the information on the joint HVGs. For the in-place version, see highly_variable_genes!

More specifically, it is the Julia re-implementation of the corresponding scanpy function For implementation details, please check the scanpy/Seurat documentations or the source code of the lower-level _highly_variable_genes_seurat_v3 function in this package. Results are almost identical to the scanpy function. The differences have been traced back to differences in the local regression for the mean-variance relationship implemented in the Loess.jl package, that differs slightly from the corresponding Python implementation.

Arguments

adata: AnnData object

layer: optional; which layer to use for calculating the HVGs. Function assumes this is a layer of counts. If layer is not provided, adata.countmatrix is used.

n_top_genes: optional; desired number of highly variable genes. Default: 2000.

batch_key: optional; key where to look for the batch indices in adata.obs. If not provided, data is treated as one batch.

span: span to use in the loess fit for the mean-variance local regression. See the Loess.jl docs for details.

verbose: whether or not to print info on current status

Returns

Returns a dictionary containing information on the highly variable genes, specifically containing the following keys is added:

highly_variable: vector of Bools indicating which genes are highly variable

highly_variable_rank: rank of the highly variable genes according to (corrected) variance

means: vector with means of each gene

variances: vector with variances of each gene

variances_norm: normalized variances of each gene

highly_variable_nbatches: if there are batches in the dataset, logs the number of batches in which each highly variable gene was actually detected as highly variable.

scVI.highly_variable_genes! — Method

highly_variable_genes!(adata::AnnData;
    layer::Union{String,Nothing} = nothing,
    n_top_genes::Int=2000,
    batch_key::Union{String,Nothing} = nothing,
    span::Float64=0.3
    )

Computes highly variable genes per batch according to the workflows on scanpy and Seurat v3 in-place.

More specifically, it is the Julia re-implementation of the corresponding scanpy function

For implementation details, please check the scanpy/Seurat documentations or the source code of the lower-level _highly_variable_genes_seurat_v3 function in this package. Results are almost identical to the scanpy function. The differences have been traced back to differences in the local regression for the mean-variance relationship implemented in the Loess.jl package, that differs slightly from the corresponding Python implementation.

Arguments

adata: AnnData object

layer: optional; which layer to use for calculating the HVGs. Function assumes this is a layer of counts. If layer is not provided, adata.countmatrix is used.

n_top_genes: optional; desired number of highly variable genes. Default: 2000.

batch_key: optional; key where to look for the batch indices in adata.obs. If not provided, data is treated as one batch.

span: span to use in the loess fit for the mean-variance local regression. See the Loess.jl docs for details.

Returns

This is the in-place version that adds an dictionary containing information on the highly variable genes directly to the adata.vars and returns the modified AnnData object. Specifically, a dictionary with the following keys is added:

highly_variable: vector of Bools indicating which genes are highly variable

highly_variable_rank: rank of the highly variable genes according to (corrected) variance

means: vector with means of each gene

variances: vector with variances of each gene

variances_norm: normalized variances of each gene

highly_variable_nbatches: if there are batches in the dataset, logs the number of batches in which each highly variable gene was actually detected as highly variable.

scVI.subset_to_hvg! — Method

subset_to_hvg!(adata::AnnData;
    layer::Union{String,Nothing} = nothing,
    n_top_genes::Int=2000,
    batch_key::Union{String,Nothing} = nothing,
    span::Float64=0.3
)

Calculates highly variable genes with highly_variable_genes! and subsets the AnnData object to the calculated HVGs. For description of input arguments, see highly_variable_genes!

Returns: adata object subset to the calculated HVGs, both in the countmatrix/layer data used for HVG calculation and in the adata.vars dictionary.

Loading built-in datasets

There are currently three built-in datasets:

the cortex data, corresponding to the cortex dataset from the scvi-tools from Zeisel et al. 2015.

the tasic data from Tasic et al. (2016), available at Gene expression Omnibus (GEO) under accession number GSE71585. Preprocessing and additional annotation according to the original manuscript; annotations are available and loaded together with the countmatrix.

the pbmc data (PBMC8k) from Zheng et al. 2017, preprocessed according to the Bioconductor workflow.

They are stored in the repository using Git LFS. After downloading and installing it as descriped on the website, the files can be downloaded after cloning the package repo with git-lfs checkout.

Cortex data

scVI.init_cortex_from_h5ad — Function

init_cortex_from_h5ad(filename::String="cortex_anndata.h5ad")

Opens a connection to the HDF5 file saved at filename that stores the corresponding Python AnnData object created and used with scvi-tools.

Reads cortex data from an AnnData object created and used with the Python scVI and saved as HDF5 file.

Information is extracted from the file with the load_cortex_from_h5ad function.

Uses this information to fill the fields of a Julia AnnData object and returns it.

scVI.load_cortex — Function

load_cortex(path::String=""; verbose::Bool=false)

Loads build-in cortex dataset. Looks for a file cortex_anndata.h5ad in the path given as input argument. This file can be downloaded by cloning the Github pacakge repository and subsequently using Git LFS to run git-lfs checkout inside the cloned package repo.

The file is the h5 export of the Python AnnData object provided as built-in cortex dataset from scvi-tools, data is from Zeisel et al. 2015.

If the file is present, the data is loaded from the Python AnnData object and stored in an analogous Julia AnnData object. This is handled by the functions init_cortex_from_h5ad and load_cortex_from_h5ad.

Alternatively, if the h5ad file is not found in the folder, the data is downloaded directly from the original authors and processed analogous to the scvi-tools processing, and subsequently stored to a Julia AnnData object. This is handled by the function init_cortex_from_url.

Returns the Julia AnnData object.

Example

julia> load_cortex()
    AnnData object with a countmatrix with 3005 cells and 1200 genes
    layers dict with the following keys: ["counts"]
    summary statistics dict with the following keys: ["n_labels", "n_vars", "n_batch", "n_continuous_covs", "n_cells", "n_proteins"]
    unique celltypes: ["interneurons", "pyramidal SS", "pyramidal CA1", "oligodendrocytes", "microglia", "endothelial-mural", "astrocytes_ependymal"]
    training status: not trained

PBMC data

scVI.load_pbmc — Function

load_pbmc(path::String = joinpath(@__DIR__, "../data/"))

Loads build-in pbmc dataset from Zheng et al. 2017. Specifically, the PBMC8k version is used and has been preprocessed according to the Bioconductor workflow.

The function loads the following files from the folder passed as path (default: data subfolder of scVI repo.)

PBMC_counts.csv: countmatrix

PBMC_annotation.csv: cell type annotation

This file can be downloaded by cloning the package Github repository and subsequently using Git LFS to run git-lfs checkout inside the cloned package repo.

From these input files, a Julia AnnData object is created. The countmatrix contains information on cell barcodes and gene names. The gene name and celltype information is stored in the vars and obs dictionaries of the AnnData object, respectively.

Returns the Julia AnnData object.

Example

julia> load_pbmc()
    AnnData object with a countmatrix with 7480 cells and 200 genes
    unique celltypes: ["B-cells", "CD4+ T-cells", "Monocytes", "CD8+ T-cells", "NK cells", "NA", "HSC", "Erythrocytes"]
    training status: not trained

Tasic data

scVI.load_tasic — Function

load_tasic(path::String = joinpath(@__DIR__, "../data/"))

Loads build-in tasic dataset from Tasic et al. (2016).

Loads the following files from the folder passed as path (default: data subfolder of scVI repo.)

Tasic_countmat.txt: countmatrix

Tasic_celltypes.txt: cell types

Tasic_genenames.txt: gene names

Tasic_receptorandmarkers.txt: List of receptor and marker genes

This file can be downloaded by cloning the package Github repository and subsequently using Git LFS to run git-lfs checkout inside the cloned package repo. The original data is available at Gene expression Omnibus (GEO) under accession number GSE71585. Preprocessing and annotation has been prepared according to the original manuscript.

From these input files, a Julia AnnData object is created. The list of receptor and marker genes is used to annotate cells as neural vs. non-neural, and annotate the neural cells as GABA- or Glutamatergic.

These annotations together with the cell type information and the gene names and receptor/marker list are stored in Dictionaries in the obs and vars fields of the AnnData obejct.

Additionally, size factors are calculated and used for normalizing the counts. The normalized counts are stored in an additional layer named normalized_counts.

Returns the Julia AnnData object.

Example

julia> load_tasic()
    AnnData object with a countmatrix with 1679 cells and 15119 genes
    layers dict with the following keys: ["normalized_counts", "counts"]
    unique celltypes: ["Vip", "L4", "L2/3", "L2", "Pvalb", "Ndnf", "L5a", "SMC", "Astro", "L5", "Micro", "Endo", "Sst", "L6b", "Sncg", "Igtp", "Oligo", "Smad3", "OPC", "L5b", "L6a"]
    training status: not trained

scVI.subset_tasic! — Method

subset_tasic!(adata::AnnData)

Subsets an input AnnData object initialized from the Tasic data according load_tasic to the neural cells and the receptor and marker genes provided as annotation.

Specifically, the count matrix and the normalized count matrix are subset to these cells and genes, and the dictionaries with information about cells and genes in adata.obs and adata.vars are also subset accordingly.

Returns the modified AnnData object.