The scvis model

scVAE model with additional tSNE loss term, based on the scvis model proposed in Ding J, Condon A and Shah SP Interpretable dimensionality reduction of single cell transcriptome data with deep generative models. Nat Commun 9, 2002 (2018).

From the abstract:

scvis is a statistical model that captures and visualizes the low-dimensional structures in single-cell gene expression data. It is robust to the number of data points and learns a probabilistic parametric mapping function to add new data points to an existing embedding.

The additional t-SNE loss component

The model is based on adding a t-SNE objective to the VAE loss function (i.e., the ELBO). This component is supposed to help structure the latent representation to look more like t-SNE.

Similar to the usual t-SNE objective, a matrix transition of transition probabilities has to be calculated. For this, individual perplexities have to be calculated.

scVI.scvis.compute_transition_probs — Function

Computes the transition probabilities between datapoints in X using the perplexity based similarity approach.

Parameters:

X: The data matrix with shape obs x vars.

perplexity: The perplexity value to use for computing the transition probabilities. Default is 30.

Returns:

A matrix of transition probabilities with shape obs x obs, where P[i,j] represents the transition probability from datapoint i to j.

scVI.scvis.compute_differentiable_transition_probs — Function

Compute differentiable transition probabilities from a given matrix X.
Parameters:
-----------
X : AbstractMatrix{S}
    The matrix from which to compute the differentiable transition probabilities.
perplexity : S
    The perplexity to use for the computation of the transition probabilities.
Returns:
--------
P : matrix of type S
    The differentiable transition probabilities.

Based on these transition probabilities, the t-SNE loss component is calculated and integrated with the standard ELBO.

scVI.scvis.tsne_repel — Function

Compute the t-SNE cost function with repulsion term.

Args:

z: matrix of shape (latent_dim, batchsize) representing the low-dimensional embedding of the data.

P: matrix of shape (batchsize, batchsize) representing the symmetric pairwise affinities of the data points.

Returns:

The t-SNE cost function value normalized by the batchsize.

scVI.scvis.tsne_loss — Function

Computes the t-SNE loss for the given scVAE model, tsne network, input data, and transition probabilities.

Parameters: - m: scVAE model - tsne_net: Dense network for t-SNE - x: Input data, shape obs x vars - P: Transition probabilities, shape obs x obs, defaults to Nothing - epoch: Current training epoch

Returns: - kl_qp: t-SNE loss

scVI.scvis.scvis_loss — Function

Computes the loss value of the scVI model.

It takes in an scVAE model, a matrix of input data (x), and an optional matrix of transition probabilities (P).
It applies the encoder to the input data to obtain a set of parameters for the approximate posterior of the latent variable (z).
It then uses these parameters to perform the reparameterization trick to sample from the approximate posterior.
It then applies the generative model to the sampled latent variable to obtain the parameters of the likelihood.
It then computes the KL divergence between the approximate posterior and the prior, and the reconstruction loss between the input data and the likelihood.
It also computes the tsne_repel loss which is the repulsion loss between the latent representations of the data.
Finally, it returns the sum of reconstruction loss, KL divergence, and tsne_repel loss normalized by the batch size.

Parameters:
- m: an instance of the scVAE model
- x: the input data, a matrix of shape obs x vars
- P: (Optional) the transition probabilities matrix, defaults to nothing
- kl_weight: (Optional) weight for the KL divergence term, defaults to 1.0
- epoch: (Optional) current training epoch, defaults to 1
Returns:
- lossval: the loss value

Parameters:
- m: an instance of the scVAE model
- x: input data of shape obs x vars
- P: (Optional) matrix of transition probabilities. If not provided, it will be computed internally.
- kl_weight: weighting factor for KL divergence. 
- epoch: The current epoch number.

Returns:
- lossval: The final loss value

There are different versions how this can be used to train scvis models.

scVI.scvis.train_scvis_model! — Function

Trains an scVAE model with the given AnnData, TrainingArgs and mode.

Parameters:

m::scVAE: the scVAE model to be trained

adata::AnnData: the AnnData object containing the dataset

training_args::TrainingArgs: the TrainingArgs object containing the training parameters such as batch size, learning rate, etc.

mode::Symbol: the training mode, can be one of the following: :alternating :alternating_diff :joint :joint_diff

Returns:

Nothing. The model is trained in place.

Raises:

A warning if an unsupported training mode is selected.

Train an scVI model in joint mode.

Parameters:

m: scVAE model to be trained

adata: AnnData object containing the data to be used for training

training_args: TrainingArgs object containing the training hyperparameters, such as weight decay, learning rate, batch size, and maximum number of epochs

::Val{:joint}: Symbol specifying the mode of training, must be :joint

Returns:

Tuple of trained scVAE model, AnnData object